Use this infrastruture for production (#52)
The most signicative change is the addition of a new parameter for handling a CDN in a different domain that the main domain. We did that because in the middle of the migration realized that the current cdn was in discourse-prod.itsre-apps.cloud, instead of being in cdn.d.m.o. This doesn't have a big impact, but for the sake of cleanliness and to reduce code complexity, we should have always the CDN using the top level domain for each environment. Let's try to get this soon.